fix: JPA Auditing 관련 수정
@EnableJpaAuditing 개별 설정 파일로 분리
- 테스트 시 JPA Auditing 적용으로 테스트 실패
- 개별 설정 파일로 분리하여 테스트 단계에서 JPA Auditing 미설정
