+++ src/main/java/com/estsoft/guesshangeul/GuessHangeulApplication.java
@@ -2,14 +2,10 @@ package com.estsoft.guesshangeul;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@EnableJpaAuditing
 @SpringBootApplication
 public class GuessHangeulApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(GuessHangeulApplication.class, args);
 	}
-
 }
