+++ src/main/java/com/estsoft/guesshangeul/entity/QuizBoard.java
@@ -43,7 +43,7 @@ public class QuizBoard {
 	@ColumnDefault("false")
 	private Boolean isDeleted;
 
-	public QuizBoard(String title, Long userId, LocalDateTime createdAt, Boolean isDeleted) {
+	public QuizBoard(String title, Long userId, Boolean isDeleted) {
 		this.title = title;
 		this.userId = userId;
 		this.createdAt = createdAt;
